[Feat] 패키지 구조 변경으로 인한 패키지 이동 및 Manager, S3 관련 Exception 추가 후 적용
