--- pothole-core/src/main/java/pothole_solution/core/pothole/PotholeRepository.java
@@ -1,6 +1,0 @@
-package pothole_solution.core.pothole;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PotholeRepository extends JpaRepository<Pothole, Long> {
-}
