--- pothole-core/src/main/java/pothole_solution/core/util/exception/GlobalExceptionHandler.java
@@ -1,18 +1,0 @@
-package pothole_solution.core.util.exception;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import pothole_solution.core.util.response.BaseResponse;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<Object> handlePotholeException(CustomException e) {
-        ExceptionStatus status = e.getStatus();
-        return ResponseEntity
-                .status(status.getHttpStatus())
-                .body(new BaseResponse<>(status));
-    }
-}
